Refactor start of processing a supp. billing batch
https://eaflood.atlassian.net/browse/WATER-3977
https://eaflood.atlassian.net/browse/WATER-3984

Both to help resolve some issues found when cleaning up our existing process billing batch service and to support changes need for handling multi-year bill runs we need to make some changes.

The key one is `ProcessBillingBatchService` is no longer appropriate because it is actually focused on process a single billing period. So, we renamed it to `ProcessBillingPeriodService` and have created a new `ProcessBillingBatchService` responsible for orchestrating the _whole_ process.

As the primary service now, we've updated the controller to call it instead of `InitiateBillingBatchService`. This resulted in use needing to make further changes in it and the services it uses.

Along the way we spotted opportunities to clean up some other things; passing in params only to find they are hard coded elsewhere, or having params named differently to how they are used within a service.
