+++ app/services/supplementary-billing/create-billing-batch.service.js
@@ -11,7 +11,7 @@ const BillingBatchModel = require('../../models/water/billing-batch.model.js')
  * Create a new billing batch
  *
  * @param {Object} regionId The regionId for the selected region
- * @param {Object} billingPeriod The billing period in the format { startDate: 2022-04-01, endDate: 2023-03-31 }
+ * @param {Object} financialYearEndings Object that contains the from and to financial year endings
  * @param {Object} options Optional params to be overridden
  * @param {String} [options.batchType=supplementary] The type of billing batch to create. Defaults to 'supplementary'
  * @param {String} [options.scheme=sroc] The applicable charging scheme. Defaults to 'sroc'
@@ -23,14 +23,15 @@ const BillingBatchModel = require('../../models/water/billing-batch.model.js')
  *
  * @returns {module:BillingBatchModel} The newly created billing batch instance with the `.region` property populated
  */
-async function go (regionId, billingPeriod, options) {
+async function go (regionId, financialYearEndings, options) {
+  const { fromFinancialYearEnding, toFinancialYearEnding } = financialYearEndings
   const optionsData = optionsDefaults(options)
 
   const billingBatch = await BillingBatchModel.query()
     .insert({
       regionId,
-      fromFinancialYearEnding: billingPeriod.endDate.getFullYear(),
-      toFinancialYearEnding: billingPeriod.endDate.getFullYear(),
+      fromFinancialYearEnding,
+      toFinancialYearEnding,
       ...optionsData
     })
     .returning('*')
