+++ app/services/supplementary-billing/initiate-billing-batch.service.js
@@ -6,13 +6,11 @@
  */
 
 const BillingBatchModel = require('../../models/water/billing-batch.model.js')
-const BillingPeriodsService = require('./billing-periods.service.js')
+
 const ChargingModuleCreateBillRunService = require('../charging-module/create-bill-run.service.js')
 const CheckLiveBillRunService = require('./check-live-bill-run.service.js')
-const CreateBillingBatchPresenter = require('../../presenters/supplementary-billing/create-billing-batch.presenter.js')
 const CreateBillingBatchService = require('./create-billing-batch.service.js')
 const CreateBillingBatchEventService = require('./create-billing-batch-event.service.js')
-const ProcessBillingBatchService = require('./process-billing-batch.service.js')
 
 /**
  * Initiate a new billing batch
@@ -20,42 +18,32 @@ const ProcessBillingBatchService = require('./process-billing-batch.service.js')
  * Initiating a new billing batch means creating both the `billing_batch` and `event` record with the appropriate data,
  * along with a bill run record in the SROC Charging Module API.
  *
- * @param {Object} billRunRequestData Validated version of the data sent in the request to create the new billing batch
+ * @param {String} regionId Id of the region the bill run is for
+ * @param {String} user Email address of the user who initiated the bill run
  *
- * @returns {Object} Details of the newly created billing batch record
+ * @returns {module:BillingBatchModel} The newly created billing batch instance
  */
-async function go (billRunRequestData) {
-  // NOTE: It will be required in the future that we cater for a range of billing periods, as changes can be back dated
-  // up to 5 years. For now though, our delivery scope is only for the 2022-2023 billing period so the final record is
-  // extracted from the `billingPeriods` array which will currently always be for the 2022-2023 billing period.
-  const billingPeriods = BillingPeriodsService.go()
-  const billingPeriod = billingPeriods[billingPeriods.length - 1]
-
-  const { region, scheme, type, user } = billRunRequestData
-
-  const financialYear = billingPeriod.endDate.getFullYear()
-  const liveBillRunExists = await CheckLiveBillRunService.go(region, financialYear)
+async function go (financialYearEndings, regionId, user) {
+  const liveBillRunExists = await CheckLiveBillRunService.go(regionId, financialYearEndings.toFinancialYearEnding)
 
   if (liveBillRunExists) {
-    throw Error(`Batch already live for region ${region}`)
+    throw Error(`Batch already live for region ${regionId}`)
   }
 
-  const chargingModuleResult = await ChargingModuleCreateBillRunService.go(region, 'sroc')
+  const chargingModuleResult = await ChargingModuleCreateBillRunService.go(regionId, 'sroc')
 
-  const billingBatchOptions = _billingBatchOptions(type, scheme, chargingModuleResult)
-  const billingBatch = await CreateBillingBatchService.go(region, billingPeriod, billingBatchOptions)
+  const billingBatchOptions = _billingBatchOptions(chargingModuleResult)
+  const billingBatch = await CreateBillingBatchService.go(regionId, financialYearEndings, billingBatchOptions)
 
   await CreateBillingBatchEventService.go(billingBatch, user)
 
-  ProcessBillingBatchService.go(billingBatch, billingPeriod)
-
-  return _response(billingBatch)
+  return billingBatch
 }
 
-function _billingBatchOptions (type, scheme, chargingModuleResult) {
+function _billingBatchOptions (chargingModuleResult) {
   const options = {
-    scheme,
-    batchType: type
+    scheme: 'sroc',
+    batchType: 'supplementary'
   }
 
   if (chargingModuleResult.succeeded) {
@@ -69,10 +57,6 @@ function _billingBatchOptions (type, scheme, chargingModuleResult) {
   options.errorCode = BillingBatchModel.errorCodes.failedToCreateBillRun
 
   return options
-}
-
-function _response (billingBatch) {
-  return CreateBillingBatchPresenter.go(billingBatch)
 }
 
 module.exports = {
