+++ test/controllers/bill-runs/bill-runs.controller.test.js
@@ -9,7 +9,7 @@ const { describe, it, beforeEach, afterEach } = exports.lab = Lab.script()
 const { expect } = Code
 
 // Things we need to stub
-const InitiateBillingBatchService = require('../../../app/services/supplementary-billing/initiate-billing-batch.service.js')
+const ProcessBillingBatchService = require('../../../app/services/supplementary-billing/process-billing-batch.service.js')
 const Boom = require('@hapi/boom')
 
 // For running our service
@@ -58,12 +58,14 @@ describe('Bill Runs controller', () => {
       }
 
       beforeEach(async () => {
-        Sinon.stub(InitiateBillingBatchService, 'go').resolves(validResponse)
+        Sinon.stub(ProcessBillingBatchService, 'go').resolves(validResponse)
       })
 
       it('returns a 200 response including details of the new billing batch', async () => {
         const response = await server.inject(options())
+        console.log('🚀 ~ file: bill-runs.controller.test.js:66 ~ it ~ response:', response)
         const payload = JSON.parse(response.payload)
+        console.log('🚀 ~ file: bill-runs.controller.test.js:68 ~ it ~ payload:', payload)
 
         expect(response.statusCode).to.equal(200)
         expect(payload).to.equal(validResponse)
@@ -84,7 +86,7 @@ describe('Bill Runs controller', () => {
       describe('because the billing batch could not be initiated', () => {
         beforeEach(async () => {
           Sinon.stub(Boom, 'badImplementation').returns(new Boom.Boom('Bang', { statusCode: 500 }))
-          Sinon.stub(InitiateBillingBatchService, 'go').rejects()
+          Sinon.stub(ProcessBillingBatchService, 'go').rejects()
         })
 
         it('returns an error response', async () => {
