+++ test/services/supplementary-billing/create-billing-batch.service.test.js
@@ -17,7 +17,7 @@ const RegionModel = require('../../../app/models/water/region.model.js')
 const CreateBillingBatchService = require('../../../app/services/supplementary-billing/create-billing-batch.service.js')
 
 describe('Create Billing Batch service', () => {
-  const billingPeriod = { startDate: new Date('2022-04-01'), endDate: new Date('2023-03-31') }
+  const financialYearEndings = { fromFinancialYearEnding: 2023, toFinancialYearEnding: 2024 }
   let region
 
   beforeEach(async () => {
@@ -28,12 +28,12 @@ describe('Create Billing Batch service', () => {
 
   describe('when the defaults are not overridden', () => {
     it('returns the new billing batch instance containing the defaults', async () => {
-      const result = await CreateBillingBatchService.go(region.regionId, billingPeriod)
+      const result = await CreateBillingBatchService.go(region.regionId, financialYearEndings)
 
       expect(result).to.be.an.instanceOf(BillingBatchModel)
 
       expect(result.fromFinancialYearEnding).to.equal(2023)
-      expect(result.toFinancialYearEnding).to.equal(2023)
+      expect(result.toFinancialYearEnding).to.equal(2024)
       expect(result.batchType).to.equal('supplementary')
       expect(result.scheme).to.equal('sroc')
       expect(result.source).to.equal('wrls')
@@ -55,12 +55,12 @@ describe('Create Billing Batch service', () => {
     const errorCode = 50
 
     it('returns the new billing batch instance containing the provided values', async () => {
-      const result = await CreateBillingBatchService.go(region.regionId, billingPeriod, { batchType, scheme, source, externalId, status, errorCode })
+      const result = await CreateBillingBatchService.go(region.regionId, financialYearEndings, { batchType, scheme, source, externalId, status, errorCode })
 
       expect(result).to.be.an.instanceOf(BillingBatchModel)
 
       expect(result.fromFinancialYearEnding).to.equal(2023)
-      expect(result.toFinancialYearEnding).to.equal(2023)
+      expect(result.toFinancialYearEnding).to.equal(2024)
       expect(result.batchType).to.equal(batchType)
       expect(result.scheme).to.equal(scheme)
       expect(result.source).to.equal(source)
@@ -78,12 +78,12 @@ describe('Create Billing Batch service', () => {
     const status = 'error'
 
     it('returns the new billing batch instance containing the provided values', async () => {
-      const result = await CreateBillingBatchService.go(region.regionId, billingPeriod, { externalId, status })
+      const result = await CreateBillingBatchService.go(region.regionId, financialYearEndings, { externalId, status })
 
       expect(result).to.be.an.instanceOf(BillingBatchModel)
 
       expect(result.fromFinancialYearEnding).to.equal(2023)
-      expect(result.toFinancialYearEnding).to.equal(2023)
+      expect(result.toFinancialYearEnding).to.equal(2024)
       expect(result.batchType).to.equal('supplementary')
       expect(result.scheme).to.equal('sroc')
       expect(result.source).to.equal('wrls')
